+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `scripts/makevideo`: Add a script to produce video of how to use tman
 - `cli/prj`: Add prompt before deleting a project
+- `cli`: Pass system config values via option
 
 ### Changed